[Fix] GunUsed, GrenadeUsed, KnifeUsed 등 이벤트 파라미터 값이 일치하지 않는 이슈 해결
1. GunUsed & GrenadeUsed는 float => int로 변경
2. KnifeUsed의 경우 이벤트 발생 시키는 코드 제거

+++ Assets/01.Scripts/UI/HUDManager.cs
@@ -49,8 +49,8 @@ namespace _01.Scripts.UI
         {
             base.Awake();
         
-            EventManager<AttackEvents>.StartListening<float>(AttackEvents.GunUsed, SetBulletCount);
-            EventManager<AttackEvents>.StartListening<float>(AttackEvents.GrenadeUsed, SetGrenadeCount);
+            EventManager<AttackEvents>.StartListening<int>(AttackEvents.GunUsed, SetBulletCount);
+            EventManager<AttackEvents>.StartListening<int>(AttackEvents.GrenadeUsed, SetGrenadeCount);
             
             EventManager<PlayerEvents>.StartListening(PlayerEvents.PlayerDead, SetBulletCountToInfinity);
             EventManager<PlayerEvents>.StartListening(PlayerEvents.PlayerDead, OnPlayerDeath);
@@ -104,12 +104,12 @@ namespace _01.Scripts.UI
             scoreGUI.text = score.ToString();
         }
 
-        private void SetGrenadeCount(float grenadeCount)
+        private void SetGrenadeCount(int grenadeCount)
         {
             grenadeCountGUI.text = grenadeCount.ToString();
         }
 
-        private void SetBulletCount(float bulletCount)
+        private void SetBulletCount(int bulletCount)
         {
             if (bulletCount > 0)
             {
